Move authorize code to separate module
To clearly distinguish it from webhooks.
